Pruebas y fracasos
Muchos fracasos

+++ Desktop App/CentralApp/src/connection/WorkerRunnable.java
@@ -110,7 +110,7 @@ public class WorkerRunnable implements Runnable {
 		String comand = null;
 		
 		
-		if (house.makeAction(Integer.valueOf(data[1]), Integer.valueOf(data[2]))) {
+		if (house.makeAction(Short.valueOf(data[1]), Short.valueOf(data[2]))) {
 			comand = "SUCCESSFUL";
 		} else {
 			comand = "WRONG";
